comments and test fixes
