+++ credentials/oauth/account_oauth_argument.go
@@ -1,0 +1,55 @@
+package oauth
+
+import (
+	"context"
+	"fmt"
+	"strings"
+)
+
+// AccountOAuthArgument is an interface that provides the necessary information
+// to authenticate using OAuth to a specific account.
+type AccountOAuthArgument interface {
+	OAuthArgument
+
+	// GetAccountHost returns the host of the account to authenticate to.
+	GetAccountHost(ctx context.Context) string
+
+	// GetAccountId returns the account ID of the account to authenticate to.
+	GetAccountId(ctx context.Context) string
+}
+
+// BasicAccountOAuthArgument is a basic implementation of the AccountOAuthArgument
+// interface that links each account with exactly one OAuth token.
+type BasicAccountOAuthArgument struct {
+	accountHost string
+	accountID   string
+}
+
+var _ AccountOAuthArgument = BasicAccountOAuthArgument{}
+
+// NewBasicAccountOAuthArgument creates a new BasicAccountOAuthArgument.
+func NewBasicAccountOAuthArgument(accountsHost, accountID string) (BasicAccountOAuthArgument, error) {
+	if !strings.HasPrefix(accountsHost, "https://") {
+		return BasicAccountOAuthArgument{}, fmt.Errorf("accountsHost must start with 'https://': %s", accountsHost)
+	}
+	if strings.HasSuffix(accountsHost, "/") {
+		return BasicAccountOAuthArgument{}, fmt.Errorf("accountsHost must not have a trailing slash: %s", accountsHost)
+	}
+	return BasicAccountOAuthArgument{accountHost: accountsHost, accountID: accountID}, nil
+}
+
+// GetAccountHost returns the host of the account to authenticate to.
+func (a BasicAccountOAuthArgument) GetAccountHost(ctx context.Context) string {
+	return a.accountHost
+}
+
+// GetAccountId returns the account ID of the account to authenticate to.
+func (a BasicAccountOAuthArgument) GetAccountId(ctx context.Context) string {
+	return a.accountID
+}
+
+// GetCacheKey returns a unique key for caching the OAuth token for the account.
+// The key is in the format "<accountHost>/oidc/accounts/<accountID>".
+func (a BasicAccountOAuthArgument) GetCacheKey(ctx context.Context) string {
+	return fmt.Sprintf("%s/oidc/accounts/%s", a.accountHost, a.accountID)
+}
