+++ credentials/oauth/oidc.go
@@ -10,6 +10,7 @@ import (
 
 var ErrOAuthNotSupported = errors.New("databricks OAuth is not supported for this host")
 
+// GetAccountOAuthEndpoints returns the OAuth endpoints for the given account.
 func GetAccountOAuthEndpoints(ctx context.Context, accountsHost, accountId string) (*OAuthAuthorizationServer, error) {
 	return &OAuthAuthorizationServer{
 		AuthorizationEndpoint: fmt.Sprintf("%s/oidc/accounts/%s/v1/authorize", accountsHost, accountId),
@@ -17,6 +18,9 @@ func GetAccountOAuthEndpoints(ctx context.Context, accountsHost, accountId strin
 	}, nil
 }
 
+// GetWorkspaceOAuthEndpoints returns the OAuth endpoints for the given workspace,
+// It queries the OIDC discovery endpoint to get the OAuth endpoints using the
+// provided ApiClient.
 func GetWorkspaceOAuthEndpoints(ctx context.Context, c *httpclient.ApiClient, host string) (*OAuthAuthorizationServer, error) {
 	oidc := fmt.Sprintf("%s/oidc/.well-known/oauth-authorization-server", host)
 	var oauthEndpoints OAuthAuthorizationServer
@@ -26,7 +30,14 @@ func GetWorkspaceOAuthEndpoints(ctx context.Context, c *httpclient.ApiClient, ho
 	return &oauthEndpoints, nil
 }
 
+// OAuthAuthorizationServer contains the OAuth endpoints for a Databricks account
+// or workspace.
 type OAuthAuthorizationServer struct {
-	AuthorizationEndpoint string `json:"authorization_endpoint"` // ../v1/authorize
-	TokenEndpoint         string `json:"token_endpoint"`         // ../v1/token
+	// AuthorizationEndpoint is the URL to redirect users to for authorization.
+	// It typically ends with /v1/authroize.
+	AuthorizationEndpoint string `json:"authorization_endpoint"`
+
+	// TokenEndpoint is the URL to exchange an authorization code for an access token.
+	// It typically ends with /v1/token.
+	TokenEndpoint string `json:"token_endpoint"`
 }
